separating out constants
