+++ perfmetrics/scripts/fio/fio_metrics.py
@@ -9,142 +9,19 @@
 
 """
 
-from dataclasses import dataclass
 import json
 import re
 import sys
-from typing import Any, Dict, List, Tuple, Callable
+from typing import Any, Dict, List, Tuple
 
+from fio import constants as consts
 from gsheet import gsheet
 
-GLOBAL_OPTS = 'global options'
-JOBS = 'jobs'
-JOB_OPTS = 'job options'
-PARAMS = 'params'
-FILESIZE = 'filesize'
-FILESIZE_KB = 'filesize_kb'
-NUMJOBS = 'numjobs'
-THREADS = 'num_threads'
-TIMESTAMP_MS = 'timestamp_ms'
-RUNTIME = 'runtime'
-RAMPTIME = 'ramp_time'
-STARTDELAY = 'startdelay'
-START_TIME = 'start_time'
-END_TIME = 'end_time'
-RW = 'rw'
-READ = 'read'
-WRITE = 'write'
-METRICS = 'metrics'
-IOPS = 'iops'
-BW_BYTES = 'bw_bytes'
-IO_BYTES = 'io_bytes'
-LAT_NS = 'lat_ns'
-MIN = 'min'
-MAX = 'max'
-MEAN = 'mean'
-PERCENTILE = 'percentile'
-P20 = '20.000000'
-P50 = '50.000000'
-P90 = '90.000000'
-P95 = '95.000000'
 
-NS_TO_S = 10**(-9)
-
-
-@dataclass(frozen=True)
-class JobParam:
-  """Dataclass for a FIO job parameter.
-
-  name: Can be any suitable value, it refers to the output dictionary key for
-  the parameter. To be used when creating parameter dict for each job.
-  json_name: Must match the FIO job specification key. Key for parameter inside 
-  'global options'/'job options' dictionary
-    Ex: For output json = {"global options": {"filesize":"50M"}, "jobs": [
-    "job options": {"rw": "read"}]}
-    `json_name` for file size will be "filesize" and that for readwrite will be
-    "rw"
-  format_param: Function returning formatted parameter value. Needed to convert
-  parameter to plottable values
-    Ex: 'filesize' is obtained as '50M', but we need to convert it to integer
-    showing size in kb in order to maintain uniformity
-  default: Default value for the parameter
-
-  """
-  name: str
-  json_name: str
-  format_param: Callable[[str], Any]
-  default: Any
-
-
-@dataclass(frozen=True)
-class JobMetric:
-  """Dataclass for a FIO job metric.
-
-  name: Can be any suitable value, it is used as key for the metric 
-  when creating metric dict for each job
-  levels: Keys for the metric inside 'read'/'write' dictionary in each job.
-  Each value in the list must match the key in the FIO output JSON
-    Ex: For job = {'read': {'iops': 123, 'latency': {'min': 0}}}
-    levels for IOPS will be ['iops'] and for min latency-> ['latency', 'min']
-  conversion: Multiplication factor to convert the metric to the desired unit
-    Ex: Extracted latency metrics are in nanoseconds, but we need them in
-    seconds for plotting. Hence conversion=10^(-9) for latency metrics.
-  """
-  name: str
-  levels: List[str]
-  conversion: float
-
-
-REQ_JOB_PARAMS = []
-# DO NOT remove the below append line
-REQ_JOB_PARAMS.append(JobParam(RW, RW, lambda val: val, 'read'))
-
-REQ_JOB_PARAMS.append(
-    JobParam(FILESIZE_KB, FILESIZE,
-             lambda val: _convert_value(val, FILESIZE_TO_KB_CONVERSION), 0))
-REQ_JOB_PARAMS.append(JobParam(THREADS, NUMJOBS, lambda val: int(val), 1))
-# append new params here
-
-REQ_JOB_METRICS = []
-REQ_JOB_METRICS.append(JobMetric(IOPS, [IOPS], 1))
-REQ_JOB_METRICS.append(JobMetric(BW_BYTES, [BW_BYTES], 1))
-REQ_JOB_METRICS.append(JobMetric(IO_BYTES, [IO_BYTES], 1))
-REQ_JOB_METRICS.append(JobMetric('lat_s_min', [LAT_NS, MIN], NS_TO_S))
-REQ_JOB_METRICS.append(JobMetric('lat_s_max', [LAT_NS, MAX], NS_TO_S))
-REQ_JOB_METRICS.append(JobMetric('lat_s_mean', [LAT_NS, MEAN], NS_TO_S))
-REQ_JOB_METRICS.extend([
-    JobMetric('lat_s_perc_20', [LAT_NS, PERCENTILE, P20], NS_TO_S),
-    JobMetric('lat_s_perc_50', [LAT_NS, PERCENTILE, P50], NS_TO_S),
-    JobMetric('lat_s_perc_90', [LAT_NS, PERCENTILE, P90], NS_TO_S),
-    JobMetric('lat_s_perc_95', [LAT_NS, PERCENTILE, P95], NS_TO_S)])
-# append new metrics here
-
-
-# Google sheet worksheet
-WORKSHEET_NAME = 'fio_metrics!'
-
-FILESIZE_TO_KB_CONVERSION = {
-    'b': 0.001,
-    'k': 1,
-    'kb': 1,
-    'm': 10**3,
-    'mb': 10**3,
-    'g': 10**6,
-    'gb': 10**6,
-    't': 10**9,
-    'tb': 10**9,
-    'p': 10**12,
-    'pb': 10**12
-}
-
-TIME_TO_MS_CONVERSION = {
-    'us': 10**(-3),
-    'ms': 1,
-    's': 1000,
-    'm': 60*1000,
-    'h': 3600*1000,
-    'd': 24*3600*1000
-}
+consts.REQ_JOB_PARAMS.append(
+    consts.JobParam(
+        consts.FILESIZE_KB, consts.FILESIZE,
+        lambda val: _convert_value(val, consts.FILESIZE_TO_KB_CONVERSION), 0))
 
 
 def _convert_value(value, conversion_dict, default_unit=''):
@@ -166,7 +43,7 @@ def _convert_value(value, conversion_dict, default_unit=''):
     conversion_dict
     ValueError: If string has no numerical part
 
-  Ex: For args value = "5s" and conversion_dict=TIME_TO_MS_CONVERSION
+  Ex: For args value = "5s" and conversion_dict=consts.TIME_TO_MS_CONVERSION
       "5s" will be converted to 5000 milliseconds and 5000 will be returned
 
   """
@@ -195,9 +72,9 @@ def _get_rw(rw_value):
 
   """
   if rw_value in ['read', 'randread']:
-    return READ
+    return consts.READ
   if rw_value in ['write', 'randwrite']:
-    return WRITE
+    return consts.WRITE
   raise ValueError('Only read/randread/write/randwrite are supported')
 
 
@@ -257,33 +134,37 @@ class FioMetrics:
     # Creating a list of just the 'rw' job parameter. Later, we will
     # loop through the jobs from the end, therefore we are creating
     # reversed rw list for easy access
-    rw_rev_list = [job_param[RW] for job_param in reversed(job_params)]
+    rw_rev_list = [job_param[consts.RW] for job_param in reversed(job_params)]
 
     global_ramptime_ms = 0
     global_startdelay_ms = 0
-    if GLOBAL_OPTS in out_json:
-      if RAMPTIME in out_json[GLOBAL_OPTS]:
-        global_ramptime_ms = _convert_value(out_json[GLOBAL_OPTS][RAMPTIME],
-                                            TIME_TO_MS_CONVERSION, 's')
-      if STARTDELAY in out_json[GLOBAL_OPTS]:
-        global_startdelay_ms = _convert_value(out_json[GLOBAL_OPTS][STARTDELAY],
-                                              TIME_TO_MS_CONVERSION, 's')
+    if consts.GLOBAL_OPTS in out_json:
+      if consts.RAMPTIME in out_json[consts.GLOBAL_OPTS]:
+        global_ramptime_ms = _convert_value(
+            out_json[consts.GLOBAL_OPTS][consts.RAMPTIME],
+            consts.TIME_TO_MS_CONVERSION, 's')
+      if consts.STARTDELAY in out_json[consts.GLOBAL_OPTS]:
+        global_startdelay_ms = _convert_value(
+            out_json[consts.GLOBAL_OPTS][consts.STARTDELAY],
+            consts.TIME_TO_MS_CONVERSION, 's')
 
     next_end_time_ms = 0
     rev_start_end_times = []
     # Looping from end since the given time is the final end time
-    for i, job in enumerate(list(reversed(out_json[JOBS]))):
+    for i, job in enumerate(list(reversed(out_json[consts.JOBS]))):
       rw = rw_rev_list[i]
       job_rw = job[_get_rw(rw)]
       ramptime_ms = 0
       startdelay_ms = 0
-      if JOB_OPTS in job:
-        if RAMPTIME in job[JOB_OPTS]:
-          ramptime_ms = _convert_value(job[JOB_OPTS][RAMPTIME],
-                                       TIME_TO_MS_CONVERSION, 's')
-        if STARTDELAY in job[JOB_OPTS]:
-          startdelay_ms = _convert_value(job[JOB_OPTS][STARTDELAY],
-                                         TIME_TO_MS_CONVERSION, 's')
+      if consts.JOB_OPTS in job:
+        if consts.RAMPTIME in job[consts.JOB_OPTS]:
+          ramptime_ms = _convert_value(job[consts.JOB_OPTS][consts.RAMPTIME],
+                                       consts.TIME_TO_MS_CONVERSION, 's')
+        if consts.STARTDELAY in job[consts.JOB_OPTS]:
+          startdelay_ms = _convert_value(
+              job[consts.JOB_OPTS][consts.STARTDELAY],
+              consts.TIME_TO_MS_CONVERSION, 's')
+
       if ramptime_ms == 0:
         ramptime_ms = global_ramptime_ms
       if startdelay_ms == 0:
@@ -291,9 +172,9 @@ class FioMetrics:
 
       # for multiple jobs, end time of one job = start time of next job
       end_time_ms = next_end_time_ms if next_end_time_ms > 0 else out_json[
-          TIMESTAMP_MS]
+          consts.TIMESTAMP_MS]
       # job start time = job end time - job runtime - ramp time
-      start_time_ms = end_time_ms - job_rw[RUNTIME] - ramptime_ms
+      start_time_ms = end_time_ms - job_rw[consts.RUNTIME] - ramptime_ms
       next_end_time_ms = start_time_ms - startdelay_ms
 
       # converting start and end time to seconds
@@ -323,54 +204,54 @@ class FioMetrics:
       Ex: out_json = {"global options": {"filesize": "50M", "numjobs": "40"},
                       "jobs":[{"job options": {"numjobs": "10"}}]
                       }
-      For REQ_JOB_PARAMS = [
-          JobParam(
-              name= FILESIZE_KB,
-              json_name= FILESIZE,
-              format_param=lambda val: _convert_value(val,
-              FILESIZE_TO_KB_CONVERSION),
-              default = 0
+      For consts.REQ_JOB_PARAMS = [
+          consts.JobParam(
+              name= RW,
+              json_name= RW,
+              format_param=lambda val: val,
+              default = 'read'
           ),
-          JobParam(
+          consts.JobParam(
               name= THREADS,
               json_name= NUMJOBS,
               format_param=lambda val: int(val),
               default = 1
           ),
-          JobParam(
-              name= RW,
-              json_name= RW,
-              format_param=lambda val: val,
-              default = 'read'
+          consts.JobParam(
+              name= FILESIZE_KB,
+              json_name= FILESIZE,
+              format_param=lambda val: _convert_value(val,
+              consts.FILESIZE_TO_KB_CONVERSION),
+              default = 0
           )
       ]
-      Extracted parameters would be [{FILESIZE_KB: 50000, THREADS: 10, RW:
-      'read'}]
+      Extracted parameters would be [{RW:'read', THREADS: 10, FILESIZE_KB:
+      50000}]
 
 
     """
     # Job parameters specified as Global options
     # Each param is formatted according to its format function before storing
     global_params = {}
-    if GLOBAL_OPTS in out_json:
-      for param in REQ_JOB_PARAMS:
+    if consts.GLOBAL_OPTS in out_json:
+      for param in consts.REQ_JOB_PARAMS:
         # If param not present in global options, default value is used
-        if param.json_name in out_json[GLOBAL_OPTS]:
+        if param.json_name in out_json[consts.GLOBAL_OPTS]:
           global_params[param.name] = param.format_param(
-              out_json[GLOBAL_OPTS][param.json_name])
+              out_json[consts.GLOBAL_OPTS][param.json_name])
         else:
           global_params[param.name] = param.default
 
     # Job parameters specified as job options overwrite global options
     params = []
-    for job in out_json[JOBS]:
+    for job in out_json[consts.JOBS]:
       curr_job_params = {}
-      if JOB_OPTS in job:
-        for param in REQ_JOB_PARAMS:
+      if consts.JOB_OPTS in job:
+        for param in consts.REQ_JOB_PARAMS:
           # If the param is not present in job options, global param is used
-          if param.json_name in job[JOB_OPTS]:
+          if param.json_name in job[consts.JOB_OPTS]:
             curr_job_params[param.name] = param.format_param(
-                job[JOB_OPTS][param.json_name])
+                job[consts.JOB_OPTS][param.json_name])
           else:
             curr_job_params[param.name] = global_params[param.name]
 
@@ -415,11 +296,11 @@ class FioMetrics:
     start_end_times = self._get_start_end_times(fio_out, job_params)
     all_jobs = []
     # Get the required metrics for every job
-    for i, job in enumerate(fio_out[JOBS]):
-      rw = job_params[i][RW]
+    for i, job in enumerate(fio_out[consts.JOBS]):
+      rw = job_params[i][consts.RW]
       job_rw = job[_get_rw(rw)]
       job_metrics = {}
-      for metric in REQ_JOB_METRICS:
+      for metric in consts.REQ_JOB_METRICS:
         val = job_rw
         """
         For metric.levels=['lat_ns', 'percentile', '20.000000']
@@ -451,10 +332,10 @@ class FioMetrics:
         continue
 
       all_jobs.append({
-          PARAMS: job_params[i],
-          START_TIME: start_time_s,
-          END_TIME: end_time_s,
-          METRICS: job_metrics
+          consts.PARAMS: job_params[i],
+          consts.START_TIME: start_time_s,
+          consts.END_TIME: end_time_s,
+          consts.METRICS: job_metrics
       })
 
     if not all_jobs:
@@ -472,16 +353,16 @@ class FioMetrics:
     values = []
     for job in jobs:
       row = []
-      for param_val in job[PARAMS].values():
+      for param_val in job[consts.PARAMS].values():
         row.append(param_val)
 
-      row.append(job[START_TIME])
-      row.append(job[END_TIME])
-      for metric_val in job[METRICS].values():
+      row.append(job[consts.START_TIME])
+      row.append(job[consts.END_TIME])
+      for metric_val in job[consts.METRICS].values():
         row.append(metric_val)
       values.append(row)
 
-    gsheet.write_to_google_sheet(WORKSHEET_NAME, values)
+    gsheet.write_to_google_sheet(consts.WORKSHEET_NAME, values)
 
   def get_metrics(self, filepath, add_to_gsheets=True) -> List[Dict[str, Any]]:
     """Returns job metrics obtained from given filepath and writes to gsheets.
